+++ app/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'com.google.gms.google-services'
     alias(libs.plugins.android.application)
     alias(libs.plugins.kotlin.android)
-
+    id 'kotlin-parcelize'
 }
 
 android {
@@ -59,11 +59,11 @@ dependencies {
 //    implementation "androidx.compose.ui:ui-graphics"
 //    implementation "androidx.compose.ui:ui-tooling-preview"
 //    implementation "androidx.compose.material3:material3"
-    implementation "androidx.navigation:navigation-compose:2.7.7"
+    implementation libs.androidx.navigation.compose.v277
 //    implementation "androidx.hilt:hilt-navigation-compose:1.2.0"
 //    implementation "com.google.dagger:hilt-android:2.49"
 //    annotationProcessor "com.google.dagger:hilt-compiler:2.49"
-    implementation "androidx.compose.material:material:1.7.6"
+    implementation libs.androidx.material
 //
 //    // Firebase
 //    implementation platform("com.google.firebase:firebase-bom:33.1.1")
