+++ gradle/libs.versions.toml
@@ -15,7 +15,9 @@ lifecycleLivedataKtx = "2.8.7"
 lifecycleViewmodelKtx = "2.8.7"
 material3 = "1.3.1"
 materialIconsExtended = "1.7.6"
+materialVersion = "1.7.6"
 navigationCompose = "2.8.5"
+navigationComposeVersion = "2.7.7"
 navigationFragment = "2.8.5"
 navigationUi = "2.8.5"
 playServicesAuth = "21.3.0"
@@ -25,7 +27,9 @@ coreKtx = "1.15.0"
 firebaseFirestore = "25.1.1"
 
 [libraries]
+androidx-material = { module = "androidx.compose.material:material", version.ref = "materialVersion" }
 androidx-material-icons-extended = { module = "androidx.compose.material:material-icons-extended", version.ref = "materialIconsExtended" }
+androidx-navigation-compose-v277 = { module = "androidx.navigation:navigation-compose", version.ref = "navigationComposeVersion" }
 coil = { module = "io.coil-kt:coil", version.ref = "coil" }
 coil-compose = { module = "io.coil-kt:coil-compose", version.ref = "coil" }
 coil-network-okhttp = { module = "io.coil-kt.coil3:coil-network-okhttp", version.ref = "coilNetworkOkhttp" }
